refactor(comp/agent/autoexit): migrate to V2 component pattern
Split into def/impl/fx/mock packages following the V2 component
convention. Update all callers to use the new autoexitfx.Module()
and autoexit.Component from def/.
