--- comp/agent/autoexit/component_mock.go
@@ -1,13 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2024-present Datadog, Inc.
-
-//go:build test
-
-package autoexit
-
-// Mock implements mock-specific methods.
-type Mock interface {
-	Component
-}
