+++ comp/agent/autoexit/fx/fx.go
@@ -3,26 +3,19 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2024-present Datadog, Inc.
 
-//go:build test
-
-package autoexitimpl
+// Package fx provides the fx module for the autoexit component.
+package fx
 
 import (
-	"go.uber.org/fx"
-
-	"github.com/DataDog/datadog-agent/comp/agent/autoexit"
+	autoexitimpl "github.com/DataDog/datadog-agent/comp/agent/autoexit/impl"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
 
-// MockModule defines the fx options for the mock component.
-func MockModule() fxutil.Module {
+// Module defines the fx options for this component.
+func Module() fxutil.Module {
 	return fxutil.Component(
-		fx.Provide(newMock),
+		fxutil.ProvideComponentConstructor(
+			autoexitimpl.NewComponent,
+		),
 	)
-}
-
-type mock struct{}
-
-func newMock() autoexit.Component {
-	return mock{}
 }
