+++ comp/agent/autoexit/impl/manager.go
