+++ comp/agent/autoexit/impl/manager_test.go
