+++ comp/agent/autoexit/mock/mock.go
@@ -1,0 +1,20 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2024-present Datadog, Inc.
+
+//go:build test
+
+// Package mock provides a mock for the autoexit component.
+package mock
+
+import (
+	"testing"
+
+	autoexit "github.com/DataDog/datadog-agent/comp/agent/autoexit/def"
+)
+
+// Mock returns a mock for the autoexit component.
+func Mock(_ *testing.T) autoexit.Component {
+	return struct{}{}
+}
